backend/pyproject.toml
[project]
name = "fundval"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "gunicorn>=25.0.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "pytest-mock>=3.15.1",
]
